pyproject.toml
[project]
name = "rag-medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "flask>=3.1.1",
    "huggingface-hub>=0.33.4",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.28",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
